share write_file utility accross code generators
