+++ crates/re_types_builder/src/codegen/cpp/mod.rs
@@ -12,6 +12,7 @@ use proc_macro2::{Ident, TokenStream};
 use quote::{format_ident, quote};
 use rayon::prelude::*;
 
+use crate::codegen::common::write_file;
 use crate::{
     codegen::AUTOGEN_WARNING, ArrowRegistry, Docs, ElementType, ObjectField, ObjectKind, Objects,
     Type,
@@ -98,10 +99,6 @@ impl CppCodeGenerator {
         folder_name: &str,
     ) -> BTreeSet<Utf8PathBuf> {
         let folder_path = self.output_path.join(folder_name);
-        std::fs::create_dir_all(&folder_path)
-            .with_context(|| format!("{folder_path:?}"))
-            .unwrap();
-
         let mut filepaths = BTreeSet::default();
 
         // Generate folder contents:
@@ -163,19 +160,6 @@ impl crate::CodeGenerator for CppCodeGenerator {
             .flatten()
             .collect()
     }
-}
-
-fn write_file(filepath: &Utf8PathBuf, code: String) {
-    if let Ok(existing) = std::fs::read_to_string(filepath) {
-        if existing == code {
-            // Don't touch the timestamp unnecessarily
-            return;
-        }
-    }
-
-    std::fs::write(filepath, code)
-        .with_context(|| format!("{filepath}"))
-        .unwrap();
 }
 
 fn generate_hpp_cpp(
