+++ crates/re_types_builder/src/codegen/python.rs
@@ -182,19 +182,7 @@ fn write_file(filepath: &Utf8PathBuf, mut source: String) {
         }
     }
 
-    if let Ok(existing) = std::fs::read_to_string(filepath) {
-        if existing == source {
-            // Don't touch the timestamp unnecessarily
-            return;
-        }
-    }
-
-    let parent_dir = filepath.parent().unwrap();
-    std::fs::create_dir_all(parent_dir)
-        .unwrap_or_else(|err| panic!("Failed to create dir {parent_dir:?}: {err}"));
-
-    std::fs::write(filepath, source)
-        .unwrap_or_else(|err| panic!("Failed to write file {filepath:?}: {err}"));
+    super::common::write_file(filepath, source);
 }
 
 fn lib_source_code(archetype_names: &[String]) -> String {
